-Merge con master - Elimminación de Factura -Modificación en atributos de la entidad Alquiler - Creación de EstadoAlquiler (Enum) -Eliminacion de fechaInicio en la request de creacion de alquiler, siempre sera el dia en que se cree -Refactorizacion de tipo de datos para precios (de double a BigDecimal)

+++ backend/src/main/java/ProyectoFinalTienda/TiendaVideojuegos/dtos/requests/InventarioItemUpdateRequest.java
@@ -2,6 +2,8 @@ package ProyectoFinalTienda.TiendaVideojuegos.dtos.requests;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -9,7 +11,7 @@ import lombok.*;
 @Builder
 public class InventarioItemUpdateRequest {
 
-    private Double precioDiario;
+    private BigDecimal precioDiario;
     private Integer stockTotal;
     private Integer stockDisponible;
 //    private Integer stockAlquilado;
